Fix: Reading SHP
Shapefiles stopped working for some reason, could not identify it exactly but suspect it has something to do with missing the GDAL library because after testing with the same Python library versions, with and without GDAL which is of that, it seemed to only work when GDAL was installed.

Updated the Dockerfile to use Conda instead of Pip as the package manager, as it simplifies the installation of GDAL considerably. On the downside, it also increases the size of the image considerably so this is something we'll want to look at optimizing in the future.